Added startInteraction(Characters player, ArrayList<> playerList)
New method that checks the positions of the players in the player list.
If a player is found, then the Battle method is started using the player
and the player found. If no one is in that position then a random
interaction is  chosen from Battle (using a random monster), Obstacle,
and Outpost.